source-postgres: Factor out generic capture logic into `sqlcapture`
This is a pretty massive change, because I'm yanking about 50% of the
discovery/capture code out of `source-postgres` and putting it into a
separate package instead.

The idea here is that there's now a `Database` interface whose methods
are the operations which will need to be implemented for each specific
SQL database we need to support, and then the `sqlcapture.Capture` struct
encapsulates all the behaviors which should be common to them all.

Currently the unit tests have been modified as lightly as possible and
still live in `source-postgres`. A later commit will try to refactor
the test suite so the same basic patterns of behavior can be applied
to all concrete implementations.